restructure Command class
