+++ src/commands/Command.ts
@@ -1,183 +1,29 @@
 import { CommandInteraction } from "discord.js";
-import { Permissions, GroupPermissions, NormalizedOverrides, NormalizedPermissions, PermissionType } from "./Permissions";
-import { RateLimits, NormalizedRateLimits, RateLimit, NormalizedRateLimit } from "./RateLimits";
+import { CommandData, NormalizedCommandDetails, SlashCommandFileData, getCommandDataFromFileData } from "./commands";
+import { NormalizedRateLimits } from "./RateLimits";
+import { NormalizedPermissions } from "./Permissions";
 
-export type CommandDetails = {
-  cooldown?: number;
-  tags?: string[];
-  broadcastable?: boolean;
-}
+export class Command implements CommandData {
+  name: string;
 
-export type NormalizedCommandDetails = {
-  cooldown: number;
-  tags: string[];
-  broadcastable: boolean;
-}
-
-export type CommandData = {
   permissions: NormalizedPermissions;
   rateLimits: NormalizedRateLimits;
   details: NormalizedCommandDetails;
-}
 
-export abstract class Command {
-  static getCommand(): any {
-    throw new Error(`/${this.name} does not have a command builder implemented`);
+  executeFunction: (interaction: CommandInteraction, broadcast: boolean) => Promise<void>;
+
+  constructor(fileData: SlashCommandFileData) {
+    const commandData = getCommandDataFromFileData(fileData);
+
+    this.name = fileData.builder.toJSON().name; // wacky and weird
+    this.executeFunction = fileData.execute;
+    
+    this.permissions = commandData.permissions;
+    this.rateLimits = commandData.rateLimits;
+    this.details = commandData.details;
   }
 
-  static getPermissions(): Permissions {
-    return;
+  async execute(interaction: CommandInteraction, broadcast: boolean): Promise<void> {
+    return await this.executeFunction(interaction, broadcast);
   }
-
-  static getRateLimits(): RateLimits {
-    return;
-  }
-
-  static async execute(interaction: CommandInteraction, preferBroadcast: boolean): Promise<void> {
-    throw new Error(`/${this.name} does not have an execute method implemented`);
-  }
-
-  static getDetails(): CommandDetails {
-    return;
-  }
-
-  static getData(): CommandData {
-    // Get the defined permissions for this command
-    const permissions = this.getPermissions();
-
-    // Normalize the permissions to arrays of PermissionEntities
-    const normalizedPermissions: NormalizedPermissions = {
-      roles: convertToNormalizedOverrides(permissions.roles),
-      channels: convertToNormalizedOverrides(permissions.channels)
-    }
-
-    // Fix the permissions to ensure validity (removing duplicates, enforcing global permissions, etc.)
-    fixNormalizedOverrides(normalizedPermissions.roles, ["role"])
-    fixNormalizedOverrides(normalizedPermissions.channels, ["channel", "category"])
-
-    // Get the defined rate limits for this command
-    const rateLimits = this.getRateLimits();
-
-    // Normalize the rate limits to arrays of PermissionEntities
-    const normalizedRateLimits: NormalizedRateLimits = {
-      limits: convertToNormalizedRateLimits(rateLimits.limits),
-      includeBotsChannel: rateLimits.includeBotsChannel ?? false
-    }
-
-    // Fix the rate limits to ensure validity (removing incorrect types, empty limits, etc.)
-    fixNormalizedRateLimits(normalizedRateLimits);
-
-    // Get the defined details for this command
-    const details = this.getDetails() ?? {};
-
-    // Enforce default values for details
-    const normalizedDetails: NormalizedCommandDetails = {
-      cooldown: details.cooldown ?? 2,
-      tags: details.tags ?? [],
-      broadcastable: details.broadcastable ?? false
-    }
-
-    return {
-      permissions: normalizedPermissions,
-      rateLimits: normalizedRateLimits,
-      details: normalizedDetails
-    }
-  }
-}
-
-function fixNormalizedOverrides(overrides: NormalizedOverrides, allowedTypes: PermissionType[]) {
-  // Remove PermissionsObjects of the incorrect type
-  overrides.allowed = overrides.allowed.filter(perm => allowedTypes.includes(perm.type));
-  overrides.denied = overrides.denied.filter(perm => allowedTypes.includes(perm.type));
-
-  // Remove duplicates
-  overrides.allowed = overrides.allowed.filter((perm, index, self) => {
-    return self.findIndex(p => p.name === perm.name) === index;
-  });
-  overrides.denied = overrides.denied.filter((perm, index, self) => {
-    return self.findIndex(p => p.name === perm.name) === index;
-  });
-
-  // Remove roles from the allowed list if they are in the denied list
-  overrides.allowed = overrides.allowed.filter(perm => {
-    return !overrides.denied.some(denied => denied.name === perm.name);
-  });
-
-  // Check if there is no global role in either list
-  if (!overrides.allowed.some(perm => perm.name === "*") && !overrides.denied.some(perm => perm.name === "*")) {
-    if (overrides.allowed.length === 0) {
-      // Add global to the allowed list if there are no allowed roles
-      overrides.allowed.push({
-        type: allowedTypes[0],
-        name: "*"
-      });
-    } else {
-      // Otherwise, add global to the denied list
-      overrides.denied.push({
-        type: allowedTypes[0],
-        name: "*"
-      });
-    }
-  }
-}
-
-function convertToNormalizedOverrides(permissions: GroupPermissions): NormalizedOverrides {
-  // If there are no permissions, return empty Overrides
-  if (!permissions) {
-    return {
-      allowed: [],
-      denied: []
-    }
-  }
-
-  // If it's an array, put it in the allowed list of new Overrides
-  if (permissions instanceof Array) {
-    return {
-      allowed: permissions,
-      denied: []
-    }
-  }
-
-  // If it's a single object, add it to the allowed list of new Overrides
-  if ("type" in permissions) {
-    return {
-      allowed: [permissions],
-      denied: []
-    }
-  }
-
-  // It now must be an Overrides object
-  // Convert the allowed and denied lists to arrays if they aren't already
-  return {
-    allowed: permissions.allowed instanceof Array ? permissions.allowed : [permissions.allowed],
-    denied: permissions.denied instanceof Array ? permissions.denied : [permissions.denied]
-  }
-}
-
-function convertToNormalizedRateLimits(rateLimits: RateLimit[]): NormalizedRateLimit[] {
-  // If there are no rate limits, return an empty array
-  if (!rateLimits) return [];
-
-  // Otherwise, convert roles in rate limits to arrays
-  const normalizedRateLimits: NormalizedRateLimit[] = rateLimits.map(limit => {
-    const newLimit: NormalizedRateLimit = {
-      roles: limit.roles instanceof Array ? limit.roles : [limit.roles],
-      window: limit.window,
-      max: limit.max
-    }
-
-    return newLimit;
-  });
-
-  return normalizedRateLimits;
-}
-
-function fixNormalizedRateLimits(rateLimits: NormalizedRateLimits) {
-  // Remove incorrect types from the roles list of each rate limit
-  rateLimits.limits.forEach(limit => {
-    limit.roles = limit.roles.filter(role => role.type === "role");
-  });
-
-  // Remove rate limits with no roles
-  rateLimits.limits = rateLimits.limits.filter(limit => limit.roles.length > 0);
 }
